make it possible to save multiple classes in the class config file. Also load these classes when the server starts, and after creating a new class or editing an existing class (using the command).
convert ArenaItem to a class because record made the values final which doesn't work well with my way of loading and saving the items.
change from using an ArrayList containing ArenaClass elements to a Hashmap containing keys being the class names, and the values being the ArenaClass. This is to avoid duplicate class entries.
make ArenaClass name transient to avoid redundancy in the config file

make new ArenaClassConfig instances when the class is needed instead of making it a static public object in MobArena.java, might reverse this if it causes too much of a mess.