Add getClosestPlayer method that takes an entity and finds out which arena player is closest to it.
Refactor executorService inside the startNextWave method so that it's called waveService and move it outside the method so that it can be shut down when an arena stops. Solves an exploit where joining very quickly again after dying would allow players to skip waves.

Refactor leaving methods leavePlayer and exitAllPlayers and move item retrieval for all players into a separate for loop from clearing and transporting to see if it solves the rare crash where the game throws an index out of bounds exception. Also changed the order in which the arena cleans things up as that too may have something to do with the crash.

Make countdown message appear on the actionbar instead of the chat.