Load custom single and recurrent waves along with custom mobs.
Add wave number requirement to if statement in pickWave so that it can only pick recurrent waves if their first wave has been reached yet.

Make countDeadMobs check if the amount of dead monsters is equal to the sum of the mobs of a wave, this makes the default counting system compatible with custom waves that have custom amounts of mobs.