Make the command that removes mob spawn points select the closest mob spawn point and delete it from the database: Make mob spawn points use BlockPos instead of Vec3i. Make the points integers instead of doubles. Get the player's distance to the various mob spawn points when using the remove command and delete the point with the lowest distance to the player. Add translatable text for showing and removing mob spawn points.
Fix incorrect formatting for the golden apple entry in the default reinforcements.
Minor refactoring in database: use one line of code where it looks more readable.
Fix a crash when trying to create the database because the player table statement was used after some tables that required the player table.